+++ zingo-netutils/src/globally_public.rs
@@ -96,11 +96,7 @@ impl TransparentIndexer for GrpcIndexer {
     ) -> Result<tonic::Streaming<RawTransaction>, tonic::Status> {
         let mut request = Request::new(filter);
         request.set_timeout(timeout);
-        Ok(self
-            .clear_net_client
-            .get_taddress_txids(request)
-            .await?
-            .into_inner())
+        Ok(self.client.get_taddress_txids(request).await?.into_inner())
     }
 
     async fn get_taddress_transactions(
@@ -111,7 +107,7 @@ impl TransparentIndexer for GrpcIndexer {
         let mut request = Request::new(filter);
         request.set_timeout(timeout);
         Ok(self
-            .clear_net_client
+            .client
             .get_taddress_transactions(request)
             .await?
             .into_inner())
@@ -125,7 +121,7 @@ impl TransparentIndexer for GrpcIndexer {
         let mut request = Request::new(addresses);
         request.set_timeout(timeout);
         Ok(self
-            .clear_net_client
+            .client
             .get_taddress_balance(request)
             .await?
             .into_inner())
@@ -137,7 +133,7 @@ impl TransparentIndexer for GrpcIndexer {
     ) -> Result<Balance, tonic::Status> {
         let stream = tokio_stream::iter(addresses);
         Ok(self
-            .clear_net_client
+            .client
             .get_taddress_balance_stream(stream)
             .await?
             .into_inner())
@@ -150,11 +146,7 @@ impl TransparentIndexer for GrpcIndexer {
     ) -> Result<GetAddressUtxosReplyList, tonic::Status> {
         let mut request = Request::new(arg);
         request.set_timeout(timeout);
-        Ok(self
-            .clear_net_client
-            .get_address_utxos(request)
-            .await?
-            .into_inner())
+        Ok(self.client.get_address_utxos(request).await?.into_inner())
     }
 
     async fn get_address_utxos_stream(
@@ -165,7 +157,7 @@ impl TransparentIndexer for GrpcIndexer {
         let mut request = Request::new(arg);
         request.set_timeout(timeout);
         Ok(self
-            .clear_net_client
+            .client
             .get_address_utxos_stream(request)
             .await?
             .into_inner())